separated setup_ddp function as it is used multple places
